dermaid/requirements.txt
torch
torchvision
albumentations
opencv-python
imbalanced-learn
scikit-learn
numpy
pandas
matplotlib
seaborn
pytorch-grad-cam
onnx
onnx-tf
tensorflow
tflite-runtime
fastapi
uvicorn
pillow
tqdm
wandb
